Add tests for `onarg` callback.
