+++ clink/lua/test/args.cpp
@@ -891,424 +891,5 @@ TEST_CASE("Lua arg parsers")
         }
     }
 
-    SECTION("Adaptive")
-    {
-        const char* script = "\
-            local generation = 0\
-            local reset = true\
-            \
-            clink.onbeginedit(function ()\
-                generation = generation + 1\
-            end)\
-            \
-            clink.onendedit(function (line)\
-                if line:find('reset') then\
-                    reset = true\
-                end\
-            end)\
-            \
-            local function ondelayinitarg(matcher, argindex)\
-                return { 'mb'..generation..'('..argindex..')' }\
-            end\
-            \
-            local function ondelayinit(matcher)\
-                if reset then\
-                    reset = false\
-                    matcher:reset()\
-                    matcher:addarg({ 'aa'..generation, 'ab'..generation, 'zz'..generation })\
-                    matcher:addarg({ delayinit=ondelayinitarg, 'ma' })\
-                end\
-            end\
-            \
-            adapt = clink.argmatcher('adapt'):setdelayinit(ondelayinit)\
-        ";
-
-        REQUIRE_LUA_DO_STRING(lua, script);
-
-        SECTION("Delayinit argmatcher")
-        {
-            lua.send_event("onbeginedit");
-            tester.set_input("adapt \x1b*");
-            tester.set_expected_output("adapt aa1 ab1 zz1 ");
-            tester.run();
-            lua_pushlstring(lua.get_state(), "adapt", 5);
-            lua.send_event("onendedit", 1);
-
-            lua.send_event("onbeginedit");
-            tester.set_input("adapt \x1b*");
-            tester.set_expected_output("adapt aa1 ab1 zz1 ");
-            tester.run();
-            lua_pushlstring(lua.get_state(), "reset", 5);
-            lua.send_event("onendedit", 1);
-
-            lua.send_event("onbeginedit");
-            tester.set_input("adapt \x1b*");
-            tester.set_expected_output("adapt aa3 ab3 zz3 ");
-            tester.run();
-            lua_pushlstring(lua.get_state(), "adapt", 5);
-            lua.send_event("onendedit", 1);
-        }
-
-        SECTION("Delayinit argument position")
-        {
-            lua.send_event("onbeginedit");
-            tester.set_input("adapt x \x1b*");
-            tester.set_expected_output("adapt x ma mb1(2) ");
-            tester.run();
-            lua_pushlstring(lua.get_state(), "adapt", 5);
-            lua.send_event("onendedit", 1);
-
-            lua.send_event("onbeginedit");
-            tester.set_input("adapt x \x1b*");
-            tester.set_expected_output("adapt x ma mb1(2) ");
-            tester.run();
-            lua_pushlstring(lua.get_state(), "reset", 5);
-            lua.send_event("onendedit", 1);
-
-            lua.send_event("onbeginedit");
-            tester.set_input("adapt x \x1b*");
-            tester.set_expected_output("adapt x ma mb3(2) ");
-            tester.run();
-            lua_pushlstring(lua.get_state(), "adapt", 5);
-            lua.send_event("onendedit", 1);
-        }
-    }
-
-    SECTION("Chaincommand")
-    {
-        // TODO: Also fromhistory tests, since the code path is different.
-
-        static const char* dir_fs[] = {
-#ifdef FIXED_FIND_ARGMATCHER_DIRNAME
-            "cd/leaf",              // Tricksy!  And caught an issue in _has_argmatcher.
-#endif
-            "nest/leaf",
-            "nest/nest2/leaf",
-            nullptr,
-        };
-
-        fs_fixture fs(dir_fs);
-
-        const char* script = "\
-            clink.argmatcher('sudo'):addflags('-x', '/x'):chaincommand()\
-            clink.argmatcher('gsudo'):addflags('-y', '/y'):chaincommand()\
-            clink.argmatcher('plerg'):addflags('-l', '-m', '-n', '/l', '/m', '/n'):addarg('aaa', 'zzz'):nofiles()\
-            clink.argmatcher('cd'):addflags('/d'):addarg(clink.dirmatches):nofiles()\
-        ";
-
-        lua_load_script(lua, app, exec);
-
-        REQUIRE_LUA_DO_STRING(lua, script);
-
-        str<> host(os::get_shellname());
-        doskey doskey(host.c_str());
-
-        MAKE_CLEANUP([&doskey](){
-            doskey.remove_alias("xyz");
-        });
-
-        SECTION("Flag at end")
-        {
-            REQUIRE(doskey.add_alias("xyz", "sudo gsudo plerg -m $*") == true);
-
-            tester.set_input("xyz ");
-            tester.set_expected_matches("aaa", "zzz");
-            tester.run();
-
-            tester.set_input("xyz -");
-            tester.set_expected_matches("-l", "-m", "-n");
-            tester.run();
-
-            tester.set_input("xyz 123 ");
-            tester.set_expected_matches();
-            tester.run();
-        }
-
-        SECTION("Flag in middle")
-        {
-            REQUIRE(doskey.add_alias("xyz", "sudo gsudo -y plerg $*") == true);
-
-            tester.set_input("xyz ");
-            tester.set_expected_matches("aaa", "zzz");
-            tester.run();
-
-            tester.set_input("xyz -");
-            tester.set_expected_matches("-l", "-m", "-n");
-            tester.run();
-
-            tester.set_input("xyz 123 ");
-            tester.set_expected_matches();
-            tester.run();
-        }
-
-        SECTION("Slash in command")
-        {
-            tester.set_input("plerg /");
-            tester.set_expected_matches("/l", "/m", "/n");
-            tester.run();
-
-            tester.set_input("plerg/");
-            tester.set_expected_matches("/l", "/m", "/n");
-            tester.run();
-
-            tester.set_input("cd /");
-            tester.set_expected_matches("/d");
-            tester.run();
-
-            tester.set_input("cd/");
-            tester.set_expected_matches("/d");
-            tester.run();
-
-            tester.set_input("sudo plerg /");
-            tester.set_expected_matches("/l", "/m", "/n");
-            tester.run();
-
-            tester.set_input("sudo plerg/");
-            tester.set_expected_matches("/l", "/m", "/n");
-            tester.run();
-
-            tester.set_input("sudo cd /");
-            tester.set_expected_matches("/d");
-            tester.run();
-
-            tester.set_input("sudo cd/");
-            tester.set_expected_matches("/d");
-            tester.run();
-
-            tester.set_input("sudo gsudo cd /");
-            tester.set_expected_matches("/d");
-            tester.run();
-
-            tester.set_input("sudo gsudo cd/");
-            tester.set_expected_matches("/d");
-            tester.run();
-        }
-
-        SECTION("Slash in word")
-        {
-#ifdef FIXED_FIND_ARGMATCHER_DIRNAME
-            tester.set_input("abcdefg cd/");
-            tester.set_expected_matches("leaf");
-            tester.run();
-#endif
-
-            tester.set_input("abcdefg nest/");
-            tester.set_expected_matches("nest\\leaf", "nest\\nest2\\");
-            tester.run();
-
-            tester.set_input("abcdefg ne/");
-            tester.set_expected_matches();
-            tester.run();
-        }
-
-        SECTION("No expand doskey")
-        {
-            // TODO: Ensure breaking `foo/` into `foo` and `/` doesn't allow expanding a doskey alias `foo`.
-        }
-    }
-
-    SECTION("nowordbreakchars")
-    {
-        // TODO: Also a fromhistory matcher, since the code path is slightly different.
-
-        static const char* bat_fs[] = {
-            "bat_thing.bat",
-            nullptr,
-        };
-
-        fs_fixture fs_bat(bat_fs);
-
-        const char* script = "\
-            local wonky = clink.argmatcher():addarg({nowordbreakchars='','12,34','12,35','12,42','3psych'})\
-            local wonderful = clink.argmatcher():addarg({nowordbreakchars=',','12,34','12,35','12,42','3psych'})\
-            local automatic = clink.argmatcher():addarg({'12,34','12,35','12,42','3psych'})\
-            clink.argmatcher('wt_thing'):addflags('-,', '-x'..wonky, '-y'..wonderful, '-z'..automatic):addarg('abc', '3ohnoes')\
-            clink.argmatcher('bat_thing'):addflags('-,', '-x'..wonky, '-y'..wonderful, '-z'..automatic):addarg('abc', '3ohnoes')\
-        ";
-
-        REQUIRE_LUA_DO_STRING(lua, script);
-
-        SECTION("Comma : yes")
-        {
-            tester.set_input("wt_thing -x 12,3");
-            tester.set_expected_matches("3ohnoes");
-            tester.run();
-
-            tester.set_input("wt_thing -x 12,");
-            tester.set_expected_matches("abc", "3ohnoes");
-            tester.run();
-        }
-
-        SECTION("Comma : no")
-        {
-            tester.set_input("wt_thing -y 12,3");
-            tester.set_expected_matches("12,34", "12,35");
-            tester.run();
-
-            tester.set_input("wt_thing -y 12,");
-            tester.set_expected_matches("12,34", "12,35", "12,42");
-            tester.run();
-        }
-
-        SECTION("Comma : auto exe")
-        {
-            tester.set_input("wt_thing -,");
-            tester.set_expected_matches("-,");
-            tester.run();
-
-            tester.set_input("wt_thing -z 12,3");
-            tester.set_expected_matches("3ohnoes");
-            tester.run();
-
-            tester.set_input("wt_thing -z 12,");
-            tester.set_expected_matches("abc", "3ohnoes");
-            tester.run();
-        }
-
-        SECTION("Comma : auto bat")
-        {
-            tester.set_input("bat_thing -,");
-            tester.set_expected_matches("abc", "3ohnoes");
-            tester.run();
-
-            tester.set_input("bat_thing -z 12,3");
-            tester.set_expected_matches("3ohnoes");
-            tester.run();
-
-            tester.set_input("bat_thing -z 12,");
-            tester.set_expected_matches("abc", "3ohnoes");
-            tester.run();
-        }
-    }
-
-    SECTION("Callbacks")
-    {
-        const char* script = "\
-            local function maybe_string(ai, w, word_index, line_state, _)\
-                local info = line_state:getwordinfo(word_index)\
-                if not info.quoted then\
-                    return 1\
-                end\
-            end\
-            local xargs = clink.argmatcher():addarg('aaa', 'bbb')\
-            clink.argmatcher('start'):addarg({onadvance=maybe_string}):addflags('-x'..xargs):chaincommand()\
-            clink.argmatcher('plerg'):addflags('-l', '-m', '-n'):addarg('xxx', 'yyy'):nofiles()\
-            \
-            local function maybe_chain(ai, word, word_index, line_state, user_data)\
-                if user_data.do_chain then\
-                    return -1\
-                elseif word == 'chain' then\
-                    user_data.do_chain = true\
-                    return 0\
-                elseif path.getextension(word) ~= '' then\
-                    return -1\
-                end\
-            end\
-            local numbers = clink.argmatcher():addarg('11', '22')\
-            local function maybe_link(l, ai, word)\
-                if ai == 0 and word == '-f' then\
-                    return numbers\
-                elseif ai > 0 and word == 'three' then\
-                    return false\
-                elseif ai > 0 and word == 'link' then\
-                    return numbers\
-                end\
-            end\
-            clink.argmatcher('qqq')\
-            :addarg({onadvance=maybe_chain, onlink=maybe_link, 'one', 'two'..numbers, 'three'..numbers})\
-            :addflags({onlink=maybe_link, '-x'})\
-            :nofiles()\
-        ";
-
-        lua_load_script(lua, app, cmd);
-        lua_load_script(lua, app, dir);
-        lua_load_script(lua, app, exec);
-        lua_word_classifier lua_classifier(lua);
-        tester.get_editor()->set_classifier(lua_classifier);
-
-        MAKE_CLEANUP([](){
-            settings::find("clink.colorize_input")->set("false");
-            settings::find("color.argmatcher")->set();
-        });
-
-        settings::find("clink.colorize_input")->set("true");
-        settings::find("color.argmatcher")->set("92");
-
-        REQUIRE_LUA_DO_STRING(lua, script);
-
-        SECTION("onadvance")
-        {
-            tester.set_input("start -x ");
-            tester.set_expected_classifications("mcf", true);
-            tester.set_expected_matches("aaa", "bbb");
-            tester.run();
-
-            tester.set_input("start qq plerg f");
-            tester.set_expected_classifications("mco", true);
-            tester.set_expected_matches("file1", "file2");
-            tester.run();
-
-            tester.set_input("start \"qq\" plerg ");
-            tester.set_expected_classifications("mcomo", true);
-            tester.set_expected_matches("xxx", "yyy");
-            tester.run();
-
-            tester.set_input("start \"qq\" plerg -l ");
-            tester.set_expected_classifications("mcomof", true);
-            tester.set_expected_matches("xxx", "yyy");
-            tester.run();
-        }
-
-        SECTION("onadvance chain")
-        {
-            tester.set_input("qqq one");
-            tester.set_expected_classifications("moa", true);
-            tester.run();
-
-            tester.set_input("qqq chain cmd");
-            tester.set_expected_classifications("moomo", true);
-            tester.run();
-
-            tester.set_input("qqq cmd");
-            tester.set_expected_classifications("moo", true);
-            tester.run();
-
-            tester.set_input("qqq cmd.exe");
-            tester.set_expected_classifications("momo", true);
-            tester.run();
-        }
-
-        SECTION("onlink")
-        {
-            tester.set_input("qqq ");
-            tester.set_expected_matches("one", "two", "three");
-            tester.run();
-
-            tester.set_input("qqq one ");
-            tester.set_expected_matches();
-            tester.run();
-
-            tester.set_input("qqq two ");
-            tester.set_expected_matches("11", "22");
-            tester.run();
-
-            tester.set_input("qqq three ");
-            tester.set_expected_matches();
-            tester.run();
-
-            tester.set_input("qqq link ");
-            tester.set_expected_matches("11", "22");
-            tester.run();
-
-            tester.set_input("qqq -x ");
-            tester.set_expected_matches("one", "two", "three");
-            tester.run();
-
-            tester.set_input("qqq -f ");
-            tester.set_expected_matches("11", "22");
-            tester.run();
-        }
-    }
-
     setting->set();
 }
